Refactor package structure for RainbowCommand and update imports in Rainbow and ArmorListener
